Catkinize perception

+++ perception/point_cloud_mapping/include/point_cloud_mapping/kdtree/kdtree.h
@@ -33,6 +33,7 @@
 #ifndef _CLOUD_KDTREE_KDTREE_H_
 #define _CLOUD_KDTREE_KDTREE_H_
 
+#include <ros/ros.h>
 // ROS includes
 #include <sensor_msgs/PointCloud.h>
 #include <geometry_msgs/Point32.h>
@@ -71,15 +72,15 @@ namespace cloud_kdtree
       /** \brief Destructor for KdTree. Deletes all allocated data arrays and destroys the kd-tree structures. */
       virtual ~KdTree () { }
 
-      virtual void nearestKSearch (const geometry_msgs::Point32 &p_q, int k, std::vector<int> &k_indices, std::vector<float> &k_distances) = 0;
-      virtual void nearestKSearch (const sensor_msgs::PointCloud &points, int index, int k, std::vector<int> &k_indices, std::vector<float> &k_distances) = 0;
-      virtual void nearestKSearch (int index, int k, std::vector<int> &k_indices, std::vector<float> &k_distances) = 0;
+      virtual void nearestKSearch (const geometry_msgs::Point32 &p_q, int k, std::vector<int> &k_indices, std::vector<double> &k_distances) = 0;
+      virtual void nearestKSearch (const sensor_msgs::PointCloud &points, int index, int k, std::vector<int> &k_indices, std::vector<double> &k_distances) = 0;
+      virtual void nearestKSearch (int index, int k, std::vector<int> &k_indices, std::vector<double> &k_distances) = 0;
 
-      virtual bool radiusSearch (const geometry_msgs::Point32 &p_q, double radius, std::vector<int> &k_indices, std::vector<float> &k_distances,
+      virtual bool radiusSearch (const geometry_msgs::Point32 &p_q, double radius, std::vector<int> &k_indices, std::vector<double> &k_distances,
                                  int max_nn = INT_MAX) = 0;
-      virtual bool radiusSearch (const sensor_msgs::PointCloud &points, int index, double radius, std::vector<int> &k_indices, std::vector<float> &k_distances,
+      virtual bool radiusSearch (const sensor_msgs::PointCloud &points, int index, double radius, std::vector<int> &k_indices, std::vector<double> &k_distances,
                                  int max_nn = INT_MAX) = 0;
-      virtual bool radiusSearch (int index, double radius, std::vector<int> &k_indices, std::vector<float> &k_distances,
+      virtual bool radiusSearch (int index, double radius, std::vector<int> &k_indices, std::vector<double> &k_distances,
 				 int max_nn = INT_MAX) = 0;
