+++ perception/point_cloud_mapping/include/point_cloud_mapping/kdtree/kdtree_ann.h
@@ -117,8 +117,8 @@ namespace cloud_kdtree
         m_lock_.unlock ();
       }
 
-      virtual void nearestKSearch (const geometry_msgs::Point32 &p_q, int k, std::vector<int> &k_indices, std::vector<float> &k_distances);
-      virtual void nearestKSearch (const sensor_msgs::PointCloud &points, int index, int k, std::vector<int> &k_indices, std::vector<float> &k_distances);
+      virtual void nearestKSearch (const geometry_msgs::Point32 &p_q, int k, std::vector<int> &k_indices, std::vector<double> &k_distances);
+      virtual void nearestKSearch (const sensor_msgs::PointCloud &points, int index, int k, std::vector<int> &k_indices, std::vector<double> &k_distances);
 
       //////////////////////////////////////////////////////////////////////////////////////////////////////////////////
       /** \brief Search for k-nearest neighbors for the given query point.
@@ -128,7 +128,7 @@ namespace cloud_kdtree
        * \param k_distances the resultant point distances
        */
       virtual inline void
-        nearestKSearch (int index, int k, std::vector<int> &k_indices, std::vector<float> &k_distances)
+        nearestKSearch (int index, int k, std::vector<int> &k_indices, std::vector<double> &k_distances)
       {
         if (index >= nr_points_)
           return;
@@ -142,8 +142,8 @@ namespace cloud_kdtree
         return;
       }
 
-      virtual bool radiusSearch (const geometry_msgs::Point32 &p_q, double radius, std::vector<int> &k_indices, std::vector<float> &k_distances, int max_nn = INT_MAX);
-      virtual bool radiusSearch (const sensor_msgs::PointCloud &points, int index, double radius, std::vector<int> &k_indices, std::vector<float> &k_distances, int max_nn = INT_MAX);
+      virtual bool radiusSearch (const geometry_msgs::Point32 &p_q, double radius, std::vector<int> &k_indices, std::vector<double> &k_distances, int max_nn = INT_MAX);
+      virtual bool radiusSearch (const sensor_msgs::PointCloud &points, int index, double radius, std::vector<int> &k_indices, std::vector<double> &k_distances, int max_nn = INT_MAX);
 
       //////////////////////////////////////////////////////////////////////////////////////////////////////////////////
       /** \brief Search for all the nearest neighbors of the query point in a given radius.
@@ -154,7 +154,7 @@ namespace cloud_kdtree
        * \param max_nn if given, bounds the maximum returned neighbors to this value
        */
       virtual inline bool
-        radiusSearch (int index, double radius, std::vector<int> &k_indices, std::vector<float> &k_distances,
+        radiusSearch (int index, double radius, std::vector<int> &k_indices, std::vector<double> &k_distances,
                       int max_nn = INT_MAX)
       {
         radius *= radius;
