+++ perception/point_cloud_mapping/src/cloud_kdtree/kdtree_ann.cpp
@@ -42,7 +42,7 @@ namespace cloud_kdtree
    * \param k_distances the resultant point distances
    */
   void
-    KdTreeANN::nearestKSearch (const geometry_msgs::Point32 &p_q, int k, std::vector<int> &k_indices, std::vector<float> &k_distances)
+    KdTreeANN::nearestKSearch (const geometry_msgs::Point32 &p_q, int k, std::vector<int> &k_indices, std::vector<double> &k_distances)
   {
     k_indices.resize (k);
     k_distances.resize (k);
@@ -67,7 +67,7 @@ namespace cloud_kdtree
    * \param k_distances the resultant point distances
    */
   void
-    KdTreeANN::nearestKSearch (const sensor_msgs::PointCloud &points, int index, int k, std::vector<int> &k_indices, std::vector<float> &k_distances)
+    KdTreeANN::nearestKSearch (const sensor_msgs::PointCloud &points, int index, int k, std::vector<int> &k_indices, std::vector<double> &k_distances)
   {
     if (index >= (int)points.points.size ())
       return;
@@ -95,7 +95,7 @@ namespace cloud_kdtree
    * \param max_nn if given, bounds the maximum returned neighbors to this value
    */
   bool
-    KdTreeANN::radiusSearch (const geometry_msgs::Point32 &p_q, double radius, std::vector<int> &k_indices, std::vector<float> &k_distances,
+    KdTreeANN::radiusSearch (const geometry_msgs::Point32 &p_q, double radius, std::vector<int> &k_indices, std::vector<double> &k_distances,
                              int max_nn)
   {
     ANNpoint p = annAllocPt (3);
@@ -135,7 +135,7 @@ namespace cloud_kdtree
    * \param max_nn if given, bounds the maximum returned neighbors to this value
    */
   bool
-    KdTreeANN::radiusSearch (const sensor_msgs::PointCloud &points, int index, double radius, std::vector<int> &k_indices, std::vector<float> &k_distances,
+    KdTreeANN::radiusSearch (const sensor_msgs::PointCloud &points, int index, double radius, std::vector<int> &k_indices, std::vector<double> &k_distances,
                              int max_nn)
   {
     ANNpoint p = annAllocPt (3);
